requirements.txt
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
django-apscheduler==0.6.0
djangorestframework-gis==0.18
django-tailwind==3.0.1
django-widget-tweaks==1.4.9
